+++ packages/valibot/src/index.ts
@@ -328,3 +328,4 @@ export class ValibotWeaver {
 
 export * from "./metadata"
 export * from "@gqloom/core"
+export * from "./types"